# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /guests/register:
    post:
      operationId: registerGuest
      tags: [guests]
      summary: Register a guest device for push notifications
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterGuestRequest"
      responses:
        "201":
          description: Guest registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Guest"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /guests/{id}:
    delete:
      operationId: deleteGuest
      tags: [guests]
      summary: Delete a guest profile
      description: >
        Permanently removes the guest record identified by id. The caller must
        supply their push token in the request body; the server verifies it
        matches the stored token for the given id, preventing one device from
        deleting another guest's profile.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteGuestRequest"
      responses:
        "204":
          description: Guest deleted successfully
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden — push token does not match this guest
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Guest not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateGuest
      tags: [guests]
      summary: Update a guest's name and room number
      description: >
        Updates the guest profile identified by id. The caller must supply their
        push token in the request body; the server verifies it matches the stored
        token for the given id, preventing one device from altering another guest's
        details.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGuestRequest"
      responses:
        "200":
          description: Guest updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Guest"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden — push token does not match this guest
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Guest not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /guests/{id}/staff:
    delete:
      operationId: staffDeleteGuest
      tags: [guests]
      summary: Remove a guest from the list (staff only)
      description: >
        Permanently removes the guest record identified by id. Requires staff
        authentication; no push token is needed.
      security:
        - staffToken: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Guest removed successfully
        "401":
          description: Unauthorized — invalid staff token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Guest not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /guests:
    get:
      operationId: getGuests
      tags: [guests]
      summary: List all registered guests (staff only)
      parameters:
        - name: roomNumber
          in: query
          required: false
          schema:
            type: string
      security:
        - staffToken: []
      responses:
        "200":
          description: List of registered guests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Guest"
        "401":
          description: Unauthorized — invalid staff token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    staffToken:
      type: http
      scheme: bearer
  schemas:
...
    UpdateGuestRequest:
      type: object
      properties:
        pushToken:
          type: string
          description: "The guest's push token — proves ownership of this profile"
        name:
          type: string
        roomNumber:
          type: string
      required:
        - pushToken
        - name
        - roomNumber
    DeleteGuestRequest:
      type: object
      properties:
        pushToken:
          type: string
          description: "The guest's push token — proves ownership of this profile"
      required:
        - pushToken
    RegisterGuestRequest:
      type: object
      properties:
        name:
          type: string
        roomNumber:
          type: string
        pushToken:
          type: string
        webPushSubscription:
          type: string
          description: "JSON-serialised Web Push subscription object for browser/PWA push delivery"
      required:
        - name
        - roomNumber
        - pushToken
    Guest:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        roomNumber:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
          description: ISO timestamp of the last profile update; equals createdAt if never changed
      required:
        - id
        - name
        - roomNumber
        - createdAt
        - updatedAt
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error